+++ amazon/cognito/client/src/main/kotlin/org/http4k/connect/amazon/cognito/action/DeleteUserPoolDomain.kt
@@ -1,0 +1,25 @@
+package org.http4k.connect.amazon.cognito.action
+
+import dev.forkhandles.result4k.Failure
+import dev.forkhandles.result4k.Success
+import org.http4k.connect.Http4kConnectAction
+import org.http4k.connect.RemoteFailure
+import org.http4k.connect.amazon.cognito.model.CloudFrontDomain
+import org.http4k.connect.amazon.cognito.model.UserPoolId
+import org.http4k.core.Method
+import org.http4k.core.Response
+import se.ansman.kotshi.JsonSerializable
+
+@Http4kConnectAction
+@JsonSerializable
+data class DeleteUserPoolDomain(
+    val UserPoolId: UserPoolId,
+    val Domain: CloudFrontDomain
+) : CognitoAction<Unit>(Unit::class) {
+    override fun toResult(response: Response) = with(response) {
+        when {
+            status.successful -> Success(Unit)
+            else -> Failure(RemoteFailure(Method.POST, uri(), status, bodyString()))
+        }
+    }
+}