+++ amazon/cognito/fake/src/main/kotlin/org/http4k/connect/amazon/cognito/FakeCognito.kt
@@ -3,18 +3,41 @@ package org.http4k.connect.amazon.cognito
 import org.http4k.aws.AwsCredentials
 import org.http4k.chaos.ChaoticHttpHandler
 import org.http4k.chaos.start
+import org.http4k.connect.amazon.AmazonJsonFake
+import org.http4k.connect.amazon.cognito.endpoints.createResourceServer
+import org.http4k.connect.amazon.cognito.endpoints.createUserPool
+import org.http4k.connect.amazon.cognito.endpoints.createUserPoolClient
+import org.http4k.connect.amazon.cognito.endpoints.createUserPoolDomain
+import org.http4k.connect.amazon.cognito.endpoints.deleteUserPool
+import org.http4k.connect.amazon.cognito.endpoints.deleteUserPoolDomain
+import org.http4k.connect.amazon.cognito.endpoints.wellKnown
+import org.http4k.connect.amazon.cognito.oauth.CognitoOAuth
+import org.http4k.connect.amazon.core.model.AwsService
 import org.http4k.connect.amazon.core.model.Region
-import org.http4k.core.Request
-import org.http4k.core.Response
-import org.http4k.core.Status.Companion.OK
+import org.http4k.connect.storage.InMemory
+import org.http4k.connect.storage.Storage
+import org.http4k.routing.routes
 import java.time.Clock
-import java.time.Clock.systemUTC
+import java.time.Duration
 
 class FakeCognito(
-    private val clock: Clock = systemUTC()
+    pools: Storage<CognitoPool> = Storage.InMemory(),
+    clock: Clock = Clock.systemUTC(),
+    expiry: Duration = Duration.ofHours(1)
 ) : ChaoticHttpHandler() {
 
-    override val app = { _: Request -> Response(OK).body("{}") }
+    private val api = AmazonJsonFake(CognitoMoshi, AwsService.of("AWSCognitoIdentityProviderService"))
+
+    override val app = routes(
+        CognitoOAuth(pools, clock, expiry),
+        wellKnown(pools),
+        api.createUserPool(pools),
+        api.createResourceServer(pools),
+        api.createUserPoolDomain(pools),
+        api.createUserPoolClient(pools),
+        api.deleteUserPoolDomain(pools),
+        api.deleteUserPool(pools)
+    )
 
     /**
      * Convenience function to get Cognito client
@@ -25,3 +48,4 @@ class FakeCognito(
 fun main() {
     FakeCognito().start()
 }
+